Token based authentication

+++ v1/controller/TaskController.php
@@ -16,6 +16,76 @@
         exit;
     }
 
+    // begin auth script
+    if (!isset($_SERVER['HTTP_AUTHORIZATION']) || strlen($_SERVER['HTTP_AUTHORIZATION']) < 1) {
+        $response = new Response();
+        $response->setHttpStatusCode(401);
+        $response->setSuccess(false);
+        (!isset($_SERVER['HTTP_AUTHORIZATION']) ? $response->addMessage("Access token is missing from the header") : false);
+        (strlen($_SERVER['HTTP_AUTHORIZATION']) < 1 ? $response->addMessage("Access token cannot be blank") : false);
+        $response->send();
+        exit;
+    }
+
+    $accesstoken = $_SERVER[HTTP_AUTHORIZATION];
+
+    try {
+        $query = $writeDb->prepare('SELECT userid, accesstokenexpiry, useractive, loginattempts FROM tbl_sessions, tbl_users WHERE tbl_sessions.userid = tbl_users.id AND accesstoken = :accesstoken');
+        $query->bindParam(':accesstoken', $accesstoken, PDO::PARAM_STR);
+        $query->execute();
+
+        $rowCount = $query->rowCount();
+        if ($rowCount === 0) {
+            $response = new Response();
+            $response->setHttpStatusCode(401);
+            $response->setSuccess(false);
+            $response->addMessage("Invalid access token");
+            $response->send();
+            exit;
+        }
+
+        $row = $query->fetch(PDO::FETCH_ASSOC);
+        $returned_userid = $row['userid'];
+        $returned_accesstokenexpiry = $row['accesstokenexpiry'];
+        $returned_useractive = $row['useractive'];
+        $returned_loginattempts = $row['loginattempts'];
+
+        if ($returned_useractive !== 'Y') {
+            $response = new Response();
+            $response->setHttpStatusCode(401);
+            $response->setSuccess(false);
+            $response->addMessage("User account not active");
+            $response->send();
+            exit;
+        }
+
+        if ($returned_loginattempts >= 3) {
+            $response = new Response();
+            $response->setHttpStatusCode(401);
+            $response->setSuccess(false);
+            $response->addMessage("User account is currently locked out");
+            $response->send();
+            exit;
+        }
+
+        if (strtotime($returned_accesstokenexpiry) < time()) {
+            $response = new Response();
+            $response->setHttpStatusCode(401);
+            $response->setSuccess(false);
+            $response->addMessage("Access token expired");
+            $response->send();
+            exit;
+        }
+    } catch (PDOException $ex) {
+        $response = new Response();
+        $response->setHttpStatusCode(500);
+        $response->setSuccess(false);
+        $response->addMessage("There was an issue authenticating - please try again");
+        $response->send();
+        exit;
+    }
+    // end auth script
+
     if (array_key_exists("taskid", $_GET)) {
         $taskid = $_GET['taskid'];
 
@@ -30,8 +100,9 @@
 
         if ($_SERVER['REQUEST_METHOD'] === 'GET') {
             try {
-                $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE id = :taskid');
+                $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE id = :taskid AND userid = :userid');
                 $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -78,8 +149,9 @@
             }
         } elseif ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
             try {
-                $query = $writeDb->prepare('DELETE FROM tbl_tasks WHERE id = :taskid');
+                $query = $writeDb->prepare('DELETE FROM tbl_tasks WHERE id = :taskid AND userid = :userid');
                 $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -160,8 +232,9 @@
                     exit;
                 }
 
-                $query = $writeDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE id = :taskid');
+                $query = $writeDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE id = :taskid AND userid = :userid');
                 $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -178,7 +251,7 @@
                     $task = new Task($row['id'], $row['title'], $row['description'], $row['deadline'], $row['completed']);
                 }
 
-                $queryString = "UPDATE tbl_tasks SET ".$queryField." WHERE id = :taskid";
+                $queryString = "UPDATE tbl_tasks SET ".$queryField." WHERE id = :taskid AND userid = :userid";
                 $query = $writeDb->prepare($queryString);
 
                 if ($title_updated === true) {
@@ -202,6 +275,7 @@
                     $query->bindParam(':completed', $up_completed, PDO::PARAM_STR);
                 }
                 $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -214,8 +288,9 @@
                     exit;
                 }
 
-                $query = $writeDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE id = :taskid');
+                $query = $writeDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE id = :taskid  AND userid = :userid');
                 $query->bindParam(':taskid', $taskid, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -283,8 +358,9 @@
 
         if ($_SERVER['REQUEST_METHOD'] === 'GET') {
             try {
-                $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE completed = :completed');
+                $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE completed = :completed  AND userid = :userid');
                 $query->bindParam(':completed', $completed, PDO::PARAM_STR);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -343,7 +419,8 @@
 
             $limitPerPage = 20;
             try {
-                $query = $readDb->prepare('SELECT count(id) as totalNoOfTasks FROM tbl_tasks');
+                $query = $readDb->prepare('SELECT count(id) as totalNoOfTasks FROM tbl_tasks AND userid = :userid');
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $row = $query->fetch(PDO::FETCH_ASSOC);
@@ -363,7 +440,8 @@
                 }
 
                 $offset = ($page == 1 ? 0 : ($limitPerPage * ($page - 1)));
-                $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks limit :pglimit offset :offset');
+                $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE userid = :userid limit :pglimit offset :offset');
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->bindParam(':pglimit', $limitPerPage, PDO::PARAM_INT);
                 $query->bindParam(':offset', $offset, PDO::PARAM_INT);
                 $query->execute();
@@ -417,7 +495,8 @@
     } elseif (empty($_GET)) {
         if ($_SERVER['REQUEST_METHOD'] === 'GET') {
             try {
-                $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks');
+                $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE userid = :userid');
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -500,11 +579,12 @@
                 $deadline = $newTask->getDeadline();
                 $completed = $newTask->getCompleted();
 
-                $query = $writeDb->prepare('INSERT INTO tbl_tasks (title, description, deadline, completed) VALUES (:title, :description, STR_TO_DATE(:deadline, \'%d/%m/%Y %H:%i\'), :completed)');
+                $query = $writeDb->prepare('INSERT INTO tbl_tasks (title, description, deadline, completed, userid) VALUES (:title, :description, STR_TO_DATE(:deadline, \'%d/%m/%Y %H:%i\'), :completed, :userid)');
                 $query->bindParam(':title', $title, PDO::PARAM_STR);
                 $query->bindParam(':description', $description, PDO::PARAM_STR);
                 $query->bindParam(':deadline', $deadline, PDO::PARAM_STR);
                 $query->bindParam(':completed', $completed, PDO::PARAM_STR);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
@@ -518,8 +598,9 @@
                 }
 
                 $lastTaskId = $writeDb->lastInsertId();
-                $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE id = :taskid');
+                $query = $readDb->prepare('SELECT id, title, description, DATE_FORMAT(deadline, "%d/%m/%Y %H:%i") as deadline, completed FROM tbl_tasks WHERE id = :taskid AND userid = :userid');
                 $query->bindParam(':taskid', $lastTaskId, PDO::PARAM_INT);
+                $query->bindParam(':userid', $returned_userid, PDO::PARAM_INT);
                 $query->execute();
 
                 $rowCount = $query->rowCount();
